% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{AddChemDist}
\alias{AddChemDist}
\title{Add chemical distance matrices to the \code{mmo}}
\usage{
AddChemDist(mmo, cos_dir = NULL, dreams_dir = NULL, m2ds_dir = NULL)
}
\arguments{
\item{mmo}{The \code{mmo}}

\item{cos_dir}{Path to the cosine similarity CSV file from MZMine (molecular networking)}

\item{dreams_dir}{Path to the DREAMS similarity CSV file from MZMine (molecular networking)}

\item{m2ds_dir}{Path to the MS2DeepScore similarity CSV file from MZMine (molecular networking)}
}
\value{
The \code{mmo} with dissimilarity matrices added (mmo$cos.dissim, mmo$dreams.dissim, mmo$m2ds.dissim)
}
\description{
This function reads cosine, DREAMS, and MS2DeepScore molecular networking outputs from MZmine,
then transform the similarity to distance and adds the dissimilarity matrices to the \code{mmo}.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
mmo <- AddChemDist(mmo,
 cos_dir = "path/to/cosine_similarity.csv",
 dreams_dir = "path/to/dreams_similarity.csv",
 m2ds_dir = "path/to/ms2deepscore_similarity.csv"
)
\dontshow{\}) # examplesIf}
}
