% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{AddFeatureInfo}
\alias{AddFeatureInfo}
\title{Add feature_info to an existing mmo from a full_feature CSV}
\usage{
AddFeatureInfo(mmo, full_feature_csv)
}
\arguments{
\item{mmo}{An existing mmo list (will be returned with $feature_info added)}

\item{full_feature_csv}{Path to the MZmine full feature table (CSV)}
}
\value{
The same mmo with mmo$feature_info set
}
\description{
Add feature_info to an existing mmo from a full_feature CSV
}
