% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{AddSiriusAnnot}
\alias{AddSiriusAnnot}
\title{Adding annotation from SIRIUS to the \code{mmo}}
\usage{
AddSiriusAnnot(
  mmo,
  canopus_structuredir,
  canopus_formuladir,
  filter_annot = FALSE,
  filter_threshold = 0.5
)
}
\arguments{
\item{mmo}{The \code{mmo}}

\item{canopus_structuredir}{Path to the SIRIUS structure_identification.tsv file}

\item{canopus_formuladir}{Path to the SIRIUS canopus_formula_summary.tsv file}

\item{filter_annot}{Logical. If TRUE, filter the annotations by probability threshold in CANOPUS.}

\item{filter_threshold}{Numeric between 0 and 1. The probability threshold for filtering annotations.}
}
\value{
The \code{mmo} with SIRIUS annotations added
}
\description{
This function reads SIRIUS structure identification and formula summary files,
and adds the annotations to the \code{mmo}.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
mmo <- AddSiriusAnnot(mmo,
 canopus_structuredir = "path/to/structure_identification.tsv",
 canopus_formuladir = "path/to/canopus_formula_summary.tsv"
)
\dontshow{\}) # examplesIf}
}
