% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{AnovaBarPlot}
\alias{AnovaBarPlot}
\title{Generate barplots for each feature and perform ANOVA}
\usage{
AnovaBarPlot(
  mmo,
  ID_list,
  outdir,
  normalization = "None",
  filter_group = FALSE,
  group_list = NULL,
  save_output = TRUE,
  width = 6,
  height = 6,
  colors = NULL,
  posthoc = NULL
)
}
\arguments{
\item{mmo}{The \code{mmo} containing metadata and feature data}

\item{ID_list}{A list of feature IDs to analyze. Use FeatureToID() to convert feature names to IDs.}

\item{outdir}{The output directory to save the bar plots and ANOVA results}

\item{normalization}{The normalization method to use for feature data. Options are 'None', 'Log', 'Meancentered', or 'Z' (default: 'None')}

\item{filter_group}{A boolean indicating whether to filter the feature values by a specific group list (default: FALSE)}

\item{group_list}{A list of groups to filter the feature values by, if filter_group is TRUE (default: NULL)}

\item{save_output}{A logical value indicating whether to save the output plot (default: TRUE)}

\item{width}{The width of the output plot in inches (default: 6)}

\item{height}{The height of the output plot in inches (default: 6)}

\item{colors}{A vector of colors for the groups (default: NULL)}

\item{posthoc}{A character string indicating the post-hoc test to perform (default: NULL)}
}
\value{
A list containing the bar plot and the ANOVA results
}
\description{
This function generates bar plots for a specified feature across different groups in the metadata, performing ANOVA and Tukey's HSD test for post-hoc analysis.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
AnovaBarPlot(mmo, ID_list = c("ID_1", "ID_2"), outdir = "output_directory", normalization = 'Z')
AnovaBarPlot(
 mmo, ID_list = c("ID_1", "ID_2"), outdir = "output_directory", normalization = 'Z',
 filter_group = TRUE, group_list = c("Group1", "Group2")
)
\dontshow{\}) # examplesIf}
}
