% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{BootCumulRichnessAUC}
\alias{BootCumulRichnessAUC}
\title{BootCumulRichnessAUC}
\usage{
BootCumulRichnessAUC(mmo, groups, n_boot = 500)
}
\arguments{
\item{mmo}{The mmo object containing feature data and metadata}

\item{groups}{A vector of group names from the metadata to consider for cumulative richness calculation}

\item{n_boot}{The number of bootstrap iterations to perform (default: 500)}
}
\value{
A numeric vector containing the normalized AUC values from each bootstrap iteration
}
\description{
This function bootstraps the normalized area under the curve (AUC) for cumulative richness by randomizing the order of groups.
It performs multiple bootstrap iterations to estimate the distribution of normalized AUC values.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
groups <- c("Control", "Treatment1", "Treatment2")
bootstrapped_aucs <- BootCumulRichnessAUC(mmo, groups, n_boot = 500)
\dontshow{\}) # examplesIf}
}
