% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{CalcNormalizedAUC}
\alias{CalcNormalizedAUC}
\title{CalcNormalizedAUC}
\usage{
CalcNormalizedAUC(curve)
}
\arguments{
\item{curve}{A data frame containing the cumulative richness curve with columns for group index and cumulative richness}
}
\value{
The normalized AUC value
}
\description{
This function calculates the normalized area under the curve (AUC) for a cumulative richness curve.
The normalized AUC is computed by dividing the AUC by the maximum possible area, which is the product of the maximum group index and maximum cumulative richness.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
curve <- GetCumulativeRichness(mmo, group =c("Control", "Treatment1", "Treatment2"))
norm_auc <- CalcNormalizedAUC(curve)
\dontshow{\}) # examplesIf}
}
