% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{CalculateCumulativeRichness}
\alias{CalculateCumulativeRichness}
\title{CalculateCumulativeRichness}
\usage{
CalculateCumulativeRichness(mmo, groups)
}
\arguments{
\item{mmo}{The mmo object containing feature data and metadata}

\item{groups}{A vector specifying the order of groups to consider for cumulative richness calculation}
}
\value{
A data frame containing the cumulative richness for each group in the specified order, with columns for group and cumulative richness.
}
\description{
This function calculates the cumulative richness of features across groups in the metadata.
Cumulative richness is defined as the total number of unique features observed as groups are added sequentially.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
groups <- c("Control", "Treatment1", "Treatment2")
cumulative_richness <- CalculateCumulativeRichness(mmo, groups)
\dontshow{\}) # examplesIf}
}
