% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{CalculateGroupBetaDistance}
\alias{CalculateGroupBetaDistance}
\title{CalculateGroupBetaDistance}
\usage{
CalculateGroupBetaDistance(mmo, beta_div, reference_group, groups)
}
\arguments{
\item{mmo}{The \code{mmo} containing feature data and metadata}

\item{beta_div}{The beta diversity distance matrix, output of GetBetaDiversity()}

\item{reference_group}{The name of the reference group to compare against}

\item{groups}{A vector of group names from the metadata to calculate distances for}
}
\value{
A data frame containing the group names, sample names, and their corresponding beta diversity distances from the reference group.
}
\description{
This function calculates the beta diversity distance between a reference group and other groups in the metadata.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
beta_diversity <- GetBetaDiversity(mmo, method = 'Gen.Uni',
 normalization = 'None', distance = 'dreams', filter_id = FALSE)
group_distances <- CalculateGroupBetaDistance(mmo, beta_div = beta_diversity,
 reference_group = 'Control', groups = c('Control', 'Treatment1', 'Treatment2'))
\dontshow{\}) # examplesIf}
}
