% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{CanopusListEnrichmentPlot}
\alias{CanopusListEnrichmentPlot}
\title{Generate a plot for enrichment analysis of Canopus-predicted terms}
\usage{
CanopusListEnrichmentPlot(
  mmo,
  id_list,
  pthr = 0.05,
  outdir,
  height = 5,
  width = 5,
  pval = "pval",
  save_output = TRUE
)
}
\arguments{
\item{mmo}{The \code{mmo} with sirius annotation and normalized data}

\item{id_list}{A vector containing names of features to analyze}

\item{pthr}{The threshold for adjusted p-value to be considered significant (default: 0.05)}

\item{outdir}{The output file path for the enrichment plot}

\item{height}{The height of the output plot in inches (default: 5)}

\item{width}{The width of the output plot in inches (default: 5)}

\item{pval}{pvalue options-pval or fdr (default: 'pval')}

\item{save_output}{boolean, whether to save the output plot (default: TRUE)}
}
\value{
A list containing the enrichment plot and the enrichment results
}
\description{
This function generates a plot for enrichment analysis of Canopus-predicted terms,
showing fold enrichment, p-value, and subset count for each term level.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
CanopusListEnrichmentPlot(
 mmo, id_list = DAMs_up$control_vs_treatment1.up,
 pthr = 0.05, outdir = 'canopus_enrichment_plot.pdf',
 height = 5, width = 5
)
\dontshow{\}) # examplesIf}
}
