% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{FeaturePhenotypeCorrelation}
\alias{FeaturePhenotypeCorrelation}
\title{FeaturePhenotypeCorrelation}
\usage{
FeaturePhenotypeCorrelation(
  mmo,
  feature_id,
  phenotype,
  groups,
  model = "lm",
  normalization = "Z",
  outdir = "FeaturePhenotypeCorrelation",
  width = 6,
  height = 6,
  save_output = TRUE
)
}
\arguments{
\item{mmo}{The \code{mmo} with feature data and metadata}

\item{feature_id}{The id of the feature to analyze}

\item{phenotype}{The name of the phenotype  in the metadata}

\item{groups}{A vector of group names from the metadata containing phenotype data}

\item{model}{The type of regression model to use. Options are 'lmm' for linear mixed model, 'lm' for simple linear regression, or 'pearson' for Pearson correlation (default: 'lmm')}

\item{normalization}{The normalization method to use for feature data. Options are 'None', 'Log', 'Meancentered', or 'Z' (default: 'Z')}

\item{outdir}{The directory to save the output files}

\item{width}{The width of the output plot in inches (default: 6)}

\item{height}{The height of the output plot in inches (default: 6)}

\item{save_output}{A logical value indicating whether to save the output plot (default: TRUE)}
}
\value{
A list of the plot and the raw data
}
\description{
This function performs correlation analysis of a specific feature against a phenotype in the metadata.
It can use linear mixed models (LMM), simple linear regression (LM), or Pearson correlation.
The default regression line of the plot uses linear model
}
