% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{FeaturePresence}
\alias{FeaturePresence}
\title{Convert feature abundances to presence / absence}
\usage{
FeaturePresence(mmo, threshold = 1)
}
\arguments{
\item{mmo}{The \code{mmo}}

\item{threshold}{Numeric threshold for presence (default = 1).
Values > threshold are set to 1, values <= threshold or NA are set to 0.}
}
\value{
The \code{mmo} with a new presence/absence table
stored in \code{mmo$feature_presence}
}
\description{
This function converts the feature abundance matrix in a \code{mmo}
into a binary presence/absence matrix and stores it as a new component
of the \code{mmo} (\code{mmo$feature_presence}).
}
\details{
A feature is considered present (1) if its abundance is greater than
a specified threshold, and absent (0) otherwise.

This function does NOT overwrite \code{mmo$feature_data}.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
mmo <- FeaturePresence(mmo, threshold = 1)
\dontshow{\}) # examplesIf}
}
