% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{FeatureToID}
\alias{FeatureToID}
\title{Convert feature names to IDs in the \code{mmo}}
\usage{
FeatureToID(mmo, feature_names)
}
\arguments{
\item{mmo}{The \code{mmo}}

\item{feature_names}{A vector of feature names to convert}
}
\value{
A vector of feature IDs corresponding to the input feature names
}
\description{
This function converts feature names to their corresponding IDs in the \code{mmo}.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
feature_ids <- FeatureToID(mmo, feature_names = c("100.0_5.0", "150.0_10.0"))
feature_ids <- FeatureToID(mmo, feature_names = mmo$feature_data$feature[1:10])
feature_ids <- FeatureToID(mmo,
 feature_names = Glucosinolates
) # if Glucosinolates is a vector of feature names
\dontshow{\}) # examplesIf}
}
