% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{GetGroupMeans}
\alias{GetGroupMeans}
\title{Calculate group means from the \code{mmo}}
\usage{
GetGroupMeans(
  mmo,
  normalization = "None",
  filter_id = FALSE,
  id_list = NULL,
  filter_group = FALSE,
  group_list = NULL
)
}
\arguments{
\item{mmo}{The \code{mmo}}

\item{normalization}{The normalization method to use. Options are 'None', 'Log', 'Meancentered', or 'Z'}

\item{filter_id}{Boolean to filter features based on a provided list (default: FALSE)}

\item{id_list}{A vector of feature names to filter (default: NULL)}

\item{filter_group}{Boolean to filter groups based on a provided list (default: FALSE)}

\item{group_list}{A vector of group names to filter (default: NULL)}
}
\value{
A data frame containing the mean feature values for each group
}
\description{
This function calculates and returns a dataframe of mean feature values for each group in the \code{mmo}, with options for normalization and filtering.
Use SwitchGroup() to change the grouping variable before running this function.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
group_means <- GetGroupMeans(mmo, normalization = 'Log')
group_means <- GetGroupMeans(mmo,
 normalization = 'None',
 filter_id = TRUE, id_list = Glucosinolates
) # if Glucosinolates is a vector of feature names
group_means <- GetGroupMeans(mmo,
 normalization = 'Z',
 filter_group = TRUE,
 group_list = c("Control", "Treatment1")
)
group_means <- GetGroupMeans(mmo,
 normalization = 'Meancentered',
 filter_id = TRUE, id_list = Glucosinolates,
 filter_group = TRUE, group_list = c("Control", "Treatment1")
) # if Glucosinolates is a vector of feature names
\dontshow{\}) # examplesIf}
}
