% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{GetHillNumbers}
\alias{GetHillNumbers}
\title{GetHillNumbers}
\usage{
GetHillNumbers(feature, metadata, q = 0, threshold = 0)
}
\arguments{
\item{feature}{Feature table with columns: id, feature, then sample columns}

\item{metadata}{Metadata table with sample and group columns}

\item{q}{The order of the Hill number to calculate (default: 0)}

\item{threshold}{Numeric; detection threshold for presence (default: 0)}
}
\value{
A data frame containing the Hill number for each group in the metadata, with columns for group and hill number.
}
\description{
This function calculates the Hill numbers for a given \code{mmo}, normalization method, and order of the Hill number without considering feature dissimilarity.
}
