% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{GetLog2FoldChange}
\alias{GetLog2FoldChange}
\title{Calculate log2 fold change for a given control group}
\usage{
GetLog2FoldChange(group_means, control_group)
}
\arguments{
\item{group_means}{A data frame containing the mean feature values for each group}

\item{control_group}{The name of the control group to compare against}
}
\value{
A data frame with log2 fold change values for each group compared to the control group
}
\description{
This function calculates the log2 fold change for each group compared to a specified control group.
}
