% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{GetNormFeature}
\alias{GetNormFeature}
\title{Retrieve feature data from the \code{mmo}, with normalization options}
\usage{
GetNormFeature(mmo, normalization)
}
\arguments{
\item{mmo}{The \code{mmo}}

\item{normalization}{The normalization method to use. Options are 'None','PA', 'Log', 'Meancentered', 'Z', or 'Imputed'}
}
\value{
The feature data corresponding to the specified normalization method
}
\description{
This function retrieves the feature data from the \code{mmo} based on the specified normalization method.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
feature_data <- GetNormFeature(mmo, normalization = 'Log')
\dontshow{\}) # examplesIf}
}
