% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{GetPerformanceFeatureLMM}
\alias{GetPerformanceFeatureLMM}
\title{GetPerformanceFeatureLMM}
\usage{
GetPerformanceFeatureLMM(
  mmo,
  phenotype,
  groups,
  DAM.list,
  comparisons,
  normalization = "Z"
)
}
\arguments{
\item{mmo}{The \code{mmo} with feature data and metadata}

\item{phenotype}{The name of the phenotype performance in the metadata}

\item{groups}{A vector of group names from the metadata containing performance data}

\item{DAM.list}{A list of DAMs to tag features}

\item{comparisons}{A list of pairwise comparisons to add fold change columns}

\item{normalization}{The normalization method to use for feature data. Options are 'None', 'Log', 'Meancentered', or 'Z' (default: 'Z')}
}
\value{
A data frame containing regression results for each feature, including effect size, p-value, and fold change columns for specified comparisons.
}
\description{
This function performs linear mixed model analysis of all features against a phenotype performance in the metadata.
}
