% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{GetRichness}
\alias{GetRichness}
\title{GetRichness}
\usage{
GetRichness(mmo, filter_feature = FALSE, feature_list = NULL)
}
\arguments{
\item{mmo}{The mmo object containing feature data and metadata}

\item{filter_feature}{A boolean indicating whether to filter features based on a provided list (default: FALSE)}

\item{feature_list}{A list of features to include in the richness calculation if filter_feature is TRUE (default: NULL)}
}
\value{
A data frame containing the richness for each sample, with columns for sample, richness, and group.
}
\description{
This function calculates the richness of features for each sample in the mmo object.
Richness is defined as the number of non-missing features observed in each sample.
}
