% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{GetSpecializationIndex}
\alias{GetSpecializationIndex}
\title{GetSpecializationIndex}
\usage{
GetSpecializationIndex(
  mmo,
  normalization = "None",
  filter_group = FALSE,
  group_list = NULL,
  filter_id = FALSE,
  id_list = NULL
)
}
\arguments{
\item{mmo}{The \code{mmo} containing feature data and metadata}

\item{normalization}{The normalization method to use for feature data.
Options are 'None', 'Log', 'Meancentered', or 'Z' (default: 'None')}

\item{filter_group}{A boolean indicating whether to filter the feature data by a specific group list (default: FALSE)}

\item{group_list}{A list of groups to filter the feature data by, if filter_group is TRUE (default: NULL)}

\item{filter_id}{A boolean indicating whether to filter the feature data by a specific list (default: FALSE)}

\item{id_list}{A list of feature names to filter the feature data by, if filter_id is TRUE (default: NULL)}
}
\value{
A data frame containing the specialization index for each group in the metadata, with columns for group and specialization index.
}
\description{
This function calculates the specialization index for a given \code{mmo}, normalization method, and optional filtering by groups and features.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
specialization_index <- GetSpecializationIndex(mmo,
                                               normalization = 'None',
                                               filter_group = FALSE)
specialization_index <- GetSpecializationIndex(mmo,
                                               normalization = 'Z',
                                               filter_group = TRUE,
                                               group_list = c('Control', 'Treatment1'),
                                               filter_id = TRUE)
\dontshow{\}) # examplesIf}
}
