% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{NMDSplot}
\alias{NMDSplot}
\title{NMDSplot}
\usage{
NMDSplot(
  mmo,
  betadiv,
  outdir,
  width = 6,
  height = 6,
  color = NULL,
  save_output = TRUE
)
}
\arguments{
\item{mmo}{The \code{mmo} containing metadata}

\item{betadiv}{The beta diversity distance matrix, output of GetBetaDiversity()}

\item{outdir}{The outdir for the output files}

\item{width}{The width of the output NMDS plot (default: 6)}

\item{height}{The height of the output NMDS plot (default: 6)}

\item{color}{A vector of colors for the groups in the plot}

\item{save_output}{A logical value indicating whether to save the output plot (default: TRUE)}
}
\value{
A list containing the NMDS plot, the NMDS coordinates, and the PERMANOVA results
}
\description{
This function generates a Non-metric Multidimensional Scaling (NMDS) plot based on the provided beta diversity distance matrix.
It also performs PERMANOVA analysis to assess the significance of group differences and saves the results to CSV files.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
beta_diversity <- GetBetaDiversity(mmo, method = 'Gen.Uni',
 normalization = 'None', distance = 'dreams', filter_id = FALSE)
# Use method = 'bray' or 'jaccard' if you want to use just feature abundance
# without considering feature spectral dissimilarity
NMDSplot(mmo, betadiv = beta_diversity, outdir = 'output/NMDS', width = 6, height = 6)
\dontshow{\}) # examplesIf}
}
