% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{PairwiseComp}
\alias{PairwiseComp}
\title{Perform pairwise comparison between two groups in the \code{mmo}}
\usage{
PairwiseComp(mmo, group1, group2, correction = "BH")
}
\arguments{
\item{mmo}{The \code{mmo}}

\item{group1}{The name of the nominator group}

\item{group2}{The name of the denominator group}

\item{correction}{The method for multiple comparison correction. Options are 'BH', 'holm', 'bonferroni', etc. Inherits from p.adjust() (default: 'BH')}
}
\value{
The \code{mmo} with pairwise comparison results added to \code{mmo$pairwise}
}
\description{
This function performs pairwise comparison between two groups in the \code{mmo},
calculating log2 fold change and adjusted p-values for given comparison of two groups.
The function adds the results to the \code{mmo$pairwise} data frame.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
mmo <- PairwiseComp(mmo, group1 = 'Control', group2 = 'Treatment1')
\dontshow{\}) # examplesIf}
}
