% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{PCAplot}
\alias{PCAplot}
\title{Plots PCA and performs PERMANOVA}
\usage{
PCAplot(
  mmo,
  color = NULL,
  outdir = "PCA",
  normalization = "Z",
  filter_id = FALSE,
  id_list = NULL,
  filter_group = FALSE,
  group_list = NULL,
  label = TRUE,
  save_output = TRUE
)
}
\arguments{
\item{mmo}{The \code{mmo} with feature data and metadata}

\item{color}{A vector of colors for the groups in the plot. Make sure the names correspond to the group names in metadata}

\item{outdir}{The output file path for the PCA plot (default: 'PCA')}

\item{normalization}{The normalization method to use for feature data. Options are 'None', 'Log', 'Meancentered', or 'Z' (default: 'Z')}

\item{filter_id}{Boolean to filter features by id_list (default: FALSE)}

\item{id_list}{A vector of feature names to filter (default: NULL)}

\item{filter_group}{Boolean to filter groups by group_list (default: FALSE)}

\item{group_list}{A vector of group names to filter (default: NULL)}

\item{label}{Boolean to indicate whether to label points with sample names (default: TRUE)}

\item{save_output}{Boolean; if TRUE (default) write plot (.pdf) and PERMANOVA
tables using \code{outdir} as prefix. If FALSE, nothing is written.}
}
\value{
A list with elements \code{plot} (ggplot), \code{df} (raw data to generate plots),
and \code{permanova} (results from \code{permanova_stat}).
}
\description{
This function performs PCA analysis and generates a PCA plot with optional filtering of features and groups.
It also conducts PERMANOVA and saves the results to CSV files.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
PCAplot(
 mmo, color = c("Control" = "blue", "Treatment1" = "red", "Treatment2" = "green"),
 outdir = 'PCA_plot', normalization = 'None',
 filter_id = FALSE, filter_group = FALSE, label = FALSE
)
PCAplot(
 mmo, color = c("Control" = "blue", "Treatment1" = "red"),
 outdir = 'PCA_plot', normalization = 'Z',
 filter_id = TRUE, id_list = Glucosinolates,
 filter_group = TRUE, group_list = c("Control", "Treatment1"), label = TRUE
)
\dontshow{\}) # examplesIf}
}
