% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{permanova_stat}
\alias{permanova_stat}
\title{Perform PERMANOVA and pairwise comparisons}
\usage{
permanova_stat(
  data,
  metadata,
  mode,
  filter_group = FALSE,
  group_list = NULL,
  permutations = 5000
)
}
\arguments{
\item{data}{A data frame or distance matrix for PERMANOVA}

\item{metadata}{A data frame containing sample metadata, including a 'group' column}

\item{mode}{The mode of the input data: 'data' for raw data or 'distance' for a distance matrix}

\item{filter_group}{Boolean to filter groups based on a provided list (default: FALSE)}

\item{group_list}{A vector of group names to filter (default: NULL)}

\item{permutations}{The number of permutations for the PERMANOVA test (default: 5000)}
}
\value{
A list containing the PERMANOVA results, raw pairwise comparison results, and a matrix of adjusted p-values
}
\description{
This function performs PERMANOVA on the given data and metadata, with options for filtering groups.
It also conducts post-hoc pairwise comparisons and adjusts p-values for multiple testing.
}
