% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{PlotFoldchangeResistanceQuad}
\alias{PlotFoldchangeResistanceQuad}
\title{PlotFoldchangeResistanceQuad}
\usage{
PlotFoldchangeResistanceQuad(
  performance_regression,
  fold_change,
  color,
  output_dir,
  save_output = TRUE,
  width = 6,
  height = 6
)
}
\arguments{
\item{performance_regression}{The regression results data frame containing effect size, fold change, and tag. The output from GetPerformanceFeatureRegression, GetPerformanceFeatureLMM, or GetPerformanceFeatureCorrelation.}

\item{fold_change}{The name of the fold change column in the performance_regression dataframe}

\item{color}{A vector of colors for the points in the plot}

\item{output_dir}{The output file path for the quadrant plot}

\item{save_output}{A logical value indicating whether to save the output plot (default: TRUE)}

\item{width}{The width of the output plot in inches (default: 6)}

\item{height}{The height of the output plot in inches (default: 6)}
}
\value{
A list containing the quadrant plot and the performance regression data frame
}
\description{
This function plots the fold change resistance in a quadrant plot, categorizing points into quadrants based on their effect size and fold change.
It also performs a binomial test to assess the distribution of points across quadrants.
}
