% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{PlotFoldchangeResistanceRegression}
\alias{PlotFoldchangeResistanceRegression}
\title{PlotFoldchangeResistanceRegression}
\usage{
PlotFoldchangeResistanceRegression(
  performance_regression,
  fold_change,
  color,
  outdir,
  width = 6,
  height = 6,
  save_output = TRUE
)
}
\arguments{
\item{performance_regression}{The regression results data frame containing effect size, fold change, and tag. The output from GetPerformanceFeatureRegression, GetPerformanceFeatureLMM, or GetPerformanceFeatureCorrelation.}

\item{fold_change}{The name of the fold change column in the performance_regression dataframe}

\item{color}{A vector of colors for the points in the plot}

\item{outdir}{The output file path for the regression plot}

\item{width}{The width of the output plot in inches (default: 6)}

\item{height}{The height of the output plot in inches (default: 6)}

\item{save_output}{A logical value indicating whether to save the output plot (default: TRUE)}
}
\value{
A list containing the regression plot and the performance regression data frame
}
\description{
This function plots the regression results of a feature against a fold change in resistance, including
regression line, p-value, and R-squared value.
}
