% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{PlotFoldchangeResistanceRegression_t}
\alias{PlotFoldchangeResistanceRegression_t}
\title{PlotFoldchangeResistanceRegression_t}
\usage{
PlotFoldchangeResistanceRegression_t(
  performance_regression,
  fold_change,
  color,
  output_dir
)
}
\arguments{
\item{performance_regression}{The regression results data frame containing effect size, fold change, and tag. The output from GetPerformanceFeatureRegression, GetPerformanceFeatureLMM, or GetPerformanceFeatureCorrelation.}

\item{fold_change}{The name of the fold change column in the performance_regression dataframe}

\item{color}{A vector of colors for the points in the plot}

\item{output_dir}{The output file path for the regression plot}
}
\description{
This function plots the regression results of a feature against a fold change in resistance, including
regression line, p-value, and R-squared value. Transposed version of PlotFoldchangeResistanceRegression.
}
