% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{PlotNPCStackedBar}
\alias{PlotNPCStackedBar}
\title{PlotNPCStackedBar}
\usage{
PlotNPCStackedBar(
  mmo,
  group_col,
  outdir,
  width = 6,
  height = 3,
  save_output = TRUE
)
}
\arguments{
\item{mmo}{The \code{mmo} with sirius annotation and normalized data}

\item{group_col}{The column name in metadata to use for grouping samples}

\item{outdir}{The output file path for the stacked bar plot (e.g., 'NPC_stacked_bar.png')}

\item{width}{The width of the output plot}

\item{height}{The height of the output plot}

\item{save_output}{boolean, whether to save the output plot}
}
\value{
A list containing the stacked bar plot and the data used to generate it
}
\description{
This function generates a stacked bar plot showing the count of features in each group categorized by NPC_pathway.
It uses the \code{mmo} with sirius annotation and normalized data.
Make sure you don't run ReplaceZero() before using this function, as it may remove presence/absence information.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
PlotNPCStackedBar(
 mmo, group_col = 'treatment',
 outdir = 'NPC_stacked_bar.png', width = 6, height = 3
)
\dontshow{\}) # examplesIf}
}
