% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{PLSDAplot}
\alias{PLSDAplot}
\title{PLS-DA plot with feature loadings}
\usage{
PLSDAplot(
  mmo,
  color,
  topk = 10,
  outdir,
  normalization = "Z",
  filter_id = FALSE,
  id_list = NULL,
  filter_group = FALSE,
  group_list = NULL,
  save_output = TRUE
)
}
\arguments{
\item{mmo}{The \code{mmo} with feature data and metadata}

\item{color}{A vector of colors for the groups in the plot. Make sure the names correspond to the group names in metadata}

\item{topk}{The number of top features to display in the plot (default: 10)}

\item{outdir}{The output file path for the PLS-DA plot}

\item{normalization}{The normalization method to use for feature data. Options are 'None', 'Log', 'Meancentered', or 'Z' (default: 'Z')}

\item{filter_id}{Boolean to filter features by id_list (default: FALSE)}

\item{id_list}{A vector of feature names to filter (default: NULL)}

\item{filter_group}{Boolean to filter groups by group_list (default: FALSE)}

\item{group_list}{A vector of group names to filter (default: NULL)}

\item{save_output}{Boolean; if TRUE (default) write plot (.pdf) and loadings
tables using \code{outdir} as prefix. If FALSE, nothing is written.}
}
\value{
A list with elements \code{plot} (ggplot), \code{df} (raw data to generate plots),
and \code{loadings} (loadings for PLSDA).
}
\description{
PLS-DA plot with feature loadings
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
PLSDAplot(
 mmo, color = c("Control" = "blue", "Treatment1" = "red", "Treatment2" = "green"),
 topk = 10, outdir = 'PLSDA_plot.pdf', normalization = 'Z',
 filter_id = FALSE, filter_group = FALSE
)
PLSDAplot(
 mmo, color = c("Control" = "blue", "Treatment1" = "red"),
 topk = 5, outdir = 'PLSDA_plot.pdf', normalization = 'Log',
 filter_id = TRUE, id_list = Glucosinolates,
 filter_group = TRUE, group_list = c("Control", "Treatment1")
)
\dontshow{\}) # examplesIf}
}
