% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{RarefactionAUC}
\alias{RarefactionAUC}
\title{RarefactionAUC}
\usage{
RarefactionAUC(rarefied_richness, n_boot = 1000, seed = 513)
}
\arguments{
\item{rarefied_richness}{The rarefied richness object containing feature data, annotations, and pairwise comparisons}

\item{n_boot}{The number of bootstrap samples to use (default: 1000)}

\item{seed}{The seed to use for the bootstrap samples (default: 513)}
}
\value{
A list containing the rarefaction AUC for each group in the metadata, with columns for group and AUC
}
\description{
This function calculates the rarefaction AUC for a given rarefied richness table
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
RarefactionAUC(rarefied_richness, n_boot = 1000)
\dontshow{\}) # examplesIf}
}
