% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{ReorderGroups}
\alias{ReorderGroups}
\title{Reorder samples in the \code{mmo} based on group order}
\usage{
ReorderGroups(mmo, group_order)
}
\arguments{
\item{mmo}{The \code{mmo}}

\item{group_order}{A vector specifying the desired order of groups}
}
\value{
The \code{mmo} with reordered samples
}
\description{
This function reorders the samples in the \code{mmo} based on a specified group order.
The function updates the order of samples in the feature data, log-normalized data, z-score data, and mean-centered data.
Use this function before plotting heatmaps or other visualizations to ensure consistent group ordering.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
mmo <- ReorderGroups(mmo, group_order = c("Control", "Treatment1", "Treatment2"))
\dontshow{\}) # examplesIf}
}
