% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{ScreenFeaturePhenotypeCorrelation}
\alias{ScreenFeaturePhenotypeCorrelation}
\title{Screen feature-phenotype correlation}
\usage{
ScreenFeaturePhenotypeCorrelation(
  mmo,
  phenotype,
  groups,
  model = "lm",
  normalization = "None"
)
}
\arguments{
\item{mmo}{The \code{mmo} with feature data and metadata}

\item{phenotype}{The name of the phenotype in the metadata}

\item{groups}{A vector of group names from the metadata containing phenotype data}

\item{model}{The type of regression model to use. Options are 'lmm' for linear mixed model, 'lm' for simple linear regression, or 'pearson', 'spearman', 'kendall' for correlation (default: 'lm')}

\item{normalization}{The normalization method to use for feature data. Options are 'None', 'Log', 'Meancentered', or 'Z' (default: 'Z')}
}
\value{
A list of the plot and the raw data
}
\description{
Use metadata-provided variables (any phenotypes or environmental variables) to screen feature-phenotype correlation
linear model, linear mixed model (using groups as random effect), or correlation (Pearson, Spearman, Kendall) are supported
}
