% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{write_anova}
\alias{write_anova}
\title{Write ANOVA and Tukey's HSD results to a CSV file}
\usage{
write_anova(anova_data, outdir, way = "oneway")
}
\arguments{
\item{anova_data}{A list containing the results of ANOVA and Tukey's HSD test}

\item{outdir}{The output directory where the results will be saved}

\item{way}{The type of ANOVA test to perform. Options are 'oneway' or 'twoway'}
}
\description{
This function writes the results of ANOVA and Tukey's HSD test to a CSV file.
}
