% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{AddSiriusAnnot}
\alias{AddSiriusAnnot}
\title{Adding annotation from SIRIUS to the mmo object}
\usage{
AddSiriusAnnot(mmo, canopus_structuredir, canopus_formuladir)
}
\arguments{
\item{mmo}{The mmo object}

\item{canopus_structuredir}{Path to the SIRIUS structure_identification.tsv file}

\item{canopus_formuladir}{Path to the SIRIUS canopus_formula_summary.tsv file}
}
\value{
The mmo object with SIRIUS annotations added
}
\description{
This function reads SIRIUS structure identification and formula summary files,
and adds the annotations to the mmo object.
}
\examples{
mmo <- AddSiriusAnnot(mmo, 
 canopus_structuredir = "path/to/structure_identification.tsv", 
 canopus_formuladir = "path/to/canopus_formula_summary.tsv"
)
}
