% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{anova_tukey_dunnett}
\alias{anova_tukey_dunnett}
\title{Perform ANOVA and Tukey's HSD test on the mmo object}
\usage{
anova_tukey_dunnett(df, formula)
}
\arguments{
\item{df}{The data frame containing the feature data and metadata}

\item{formula}{The formula for the ANOVA test, e.g., "feature ~ group"}
}
\value{
A list containing the ANOVA results, Tukey's HSD results, Tukey's significance letters, and Dunnett's test results
}
\description{
This function performs ANOVA and Tukey's HSD test on the feature data of the mmo object,
}
