% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{AnovaBarPlot}
\alias{AnovaBarPlot}
\title{AnovaBarPlot}
\usage{
AnovaBarPlot(
  mmo,
  ID_list,
  outdir,
  normalization = "None",
  filter_group = FALSE,
  group_list = NULL
)
}
\arguments{
\item{mmo}{The mmo object containing metadata and feature data}

\item{ID_list}{A list of feature IDs to analyze}

\item{outdir}{The output directory to save the bar plots and ANOVA results}

\item{normalization}{The normalization method to use for feature data. Options are 'None', 'Log', 'Meancentered', or 'Z' (default: 'None')}

\item{filter_group}{A boolean indicating whether to filter the feature values by a specific group list (default: FALSE)}

\item{group_list}{A list of groups to filter the feature values by, if filter_group is TRUE (default: NULL)}
}
\description{
This function generates bar plots for a specified feature across different groups in the metadata, performing ANOVA and Tukey's HSD test for post-hoc analysis.
}
