% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{CalculateGroupBetaDistance}
\alias{CalculateGroupBetaDistance}
\title{CalculateGroupBetaDistance}
\usage{
CalculateGroupBetaDistance(mmo, beta_div, reference_group, groups)
}
\arguments{
\item{mmo}{The mmo object containing feature data and metadata}

\item{beta_div}{The beta diversity distance matrix}

\item{reference_group}{The name of the reference group to compare against}

\item{groups}{A vector of group names from the metadata to calculate distances for}
}
\value{
A data frame containing the group names, sample names, and their corresponding beta diversity distances from the reference group.
}
\description{
This function calculates the beta diversity distance between a reference group and other groups in the metadata.
}
