% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{CanopusAllLevelEnrichmentPlot}
\alias{CanopusAllLevelEnrichmentPlot}
\title{Generate a plot for enrichment analysis of Canopus-predicted terms across all levels}
\usage{
CanopusAllLevelEnrichmentPlot(
  mmo = mmo,
  comp.list,
  terms = "all_terms",
  term_levels = NULL,
  pthr = 0.1,
  representation = "greater",
  outdir = "enrichment",
  height = 10,
  width = 8,
  pval = "pval",
  save_output = TRUE
)
}
\arguments{
\item{mmo}{The mmo object with sirius annotation and normalized data}

\item{comp.list}{A list to analyze, where each element is a vector of feature names}

\item{terms}{The terms to analyze. Options are 'all_terms', 'NPC', 'ClassyFire', or 'custom' (default: 'all_terms')}

\item{term_levels}{list of custom term levels to use}

\item{pthr}{The threshold for adjusted p-value to be considered significant (default: 0.1)}

\item{representation}{The representation type for enrichment analysis. Options are 'greater' for overrepresentation (default: 'greater')}

\item{outdir}{The output directory for saving the plot and the enrichment results (default: 'enrichment')}

\item{height}{The height of the output plot in inches (default: 10)}

\item{width}{The width of the output plot in inches (default: 8)}

\item{pval}{pvalue options-pval or fdr (default: 'pval')}

\item{save_output}{boolean, whether to save the output plot (default: TRUE)}
}
\value{
A list of the plot and the enrichment results
}
\description{
This function generates a plot for enrichment analysis of Canopus-predicted terms across all levels,
showing fold enrichment, p-value, and subset count for each term level.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
comp.list <- list(
  comparison1 = DAMs_up$control_vs_treatment1.up,
  comparison2 = DAMs_up$control_vs_treatment2.up
)
CanopusAllLevelEnrichmentPlot(
 mmo, comp.list = comp.list, terms = 'all_terms',
 pthr = 0.1, representation = 'greater', prefix = 'enrichment_all_levels',
 height = 10, width = 8
)
CanopusAllLevelEnrichmentPlot(
 mmo, comp.list = comp.list, terms = 'NPC',
 pthr = 0.1, representation = 'greater', prefix = 'enrichment_NPC_levels',
 height = 10, width = 8
)
CanopusAllLevelEnrichmentPlot(
 mmo, comp.list = comp.list, terms = 'ClassyFire',
 pthr = 0.1, representation = 'greater', prefix = 'enrichment_ClassyFire_levels',
 height = 10, width = 8
)
\dontshow{\}) # examplesIf}
}
