% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{CanopusLevelEnrichmentAnal}
\alias{CanopusLevelEnrichmentAnal}
\title{Enrichment analysis for Canopus-predicted terms}
\usage{
CanopusLevelEnrichmentAnal(
  mmo,
  list_test,
  pthr = 0.1,
  sig = TRUE,
  term_level = "NPC_pathway",
  representation = "greater",
  pval = "pval"
)
}
\arguments{
\item{mmo}{The mmo object with sirius annotation and normalized data}

\item{list_test}{A vector containing ids of features to analyze}

\item{pthr}{The threshold for adjusted p-value to be considered significant (default: 0.1)}

\item{sig}{A logical value indicating whether to return only significant terms (default: TRUE)}

\item{term_level}{The level of term to use for enrichment analysis
Options are 'NPC_pathway', 'NPC_superclass', 'NPC_class', 'ClassyFire_superclass', 'ClassyFire_class',
'ClassyFire_subclass', 'ClassyFire_level5', or 'ClassyFire_most_specific' (default: 'NPC_pathway')}

\item{representation}{The representation type for enrichment analysis. Options are 'greater' for overrepresentation (default: 'greater')}

\item{pval}{pvalue options-pval or fdr (default: 'pval')}
}
\value{
A data frame containing the enrichment results, including term level,
term name, subset count, total count, fold enrichment, p-value, and adjusted p-value (FDR)
}
\description{
This function performs enrichment analysis for Canopus-predicted terms on a given list of features.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# Perform enrichment analysis for a list of features using NPC_pathway level
sig_terms <- CanopusLevelEnrichmentAnal(
 mmo, list_test = c("feature1", "feature2"), pthr = 0.1,
 sig = TRUE, term_level = 'NPC_pathway', representation = 'greater'
)
# Perform enrichment analysis for a list of features using
# ClassyFire_class level and return all terms
all_terms <- CanopusLevelEnrichmentAnal(
 mmo, list_test = c("feature1", "feature2"), pthr = 0.1,
 sig = FALSE, term_level = 'ClassyFire_class', representation = 'greater'
)
\dontshow{\}) # examplesIf}
}
