% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{CanopusLevelEnrichmentPlot}
\alias{CanopusLevelEnrichmentPlot}
\title{Generate a plot for enrichment analysis of Canopus-predicted terms at a specific level using a list of vectors of features}
\usage{
CanopusLevelEnrichmentPlot(
  mmo = mmo,
  comp.list,
  term_level = "NPC_pathway",
  pthr = 0.1,
  representation = "greater",
  prefix = "enrichment",
  height = 5,
  width = 5,
  pval = "pval"
)
}
\arguments{
\item{mmo}{The mmo object with sirius annotation and normalized data}

\item{comp.list}{A list to analyze, where each element is a vector of feature names}

\item{term_level}{The level of term to use for enrichment analysis.
Options are 'NPC_pathway', 'NPC_superclass', 'NPC_class',
'ClassyFire_superclass', 'ClassyFire_class', 'ClassyFire_subclass',
'ClassyFire_level5', or 'ClassyFire_most_specific' (default: 'NPC_pathway')}

\item{pthr}{The threshold for adjusted p-value to be considered significant (default: 0.1)}

\item{representation}{The representation type for enrichment analysis. Options are 'greater' for overrepresentation (default: 'greater')}

\item{prefix}{The prefix for output files (default: 'enrichment')}

\item{height}{The height of the output plot in inches (default: 5)}

\item{width}{The width of the output plot in inches (default: 5)}

\item{pval}{pvalue options-pval or fdr (default: 'pval')}
}
\description{
This function generates a plot for enrichment analysis of Canopus-predicted terms at a specific level,
showing fold enrichment, p-value, and subset count for each term.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# Perform enrichment analysis for multiple comparisons using NPC_pathway level
comp.list <- list(
  comparison1 = DAMs_up$control_vs_treatment1.up,
  comparison2 = DAMs_up$control_vs_treatment2.up
)
CanopusLevelEnrichmentPlot(
 mmo, comp.list = comp.list, term_level = 'NPC_pathway',
 pthr = 0.1, representation = 'greater', prefix = 'enrichment_plot',
 height = 5, width = 5
)
\dontshow{\}) # examplesIf}
}
