% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{ExportFeaturesToCSV}
\alias{ExportFeaturesToCSV}
\title{ExportFeaturesToCSV}
\usage{
ExportFeaturesToCSV(mmo, feature_list, normalization = "None", output_dir)
}
\arguments{
\item{mmo}{The mmo object containing feature data, annotations, and pairwise comparisons}

\item{feature_list}{A list of feature names to filter and export}

\item{normalization}{The normalization method to use for feature data. Options are 'None', 'Log', 'Meancentered', or 'Z' (default: 'None')}

\item{output_dir}{The output directory to save the CSV file}
}
\description{
This function exports selected features, their annotations, and pairwise comparisons to a CSV file.
}
