% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{FeaturePerformanceRegression}
\alias{FeaturePerformanceRegression}
\title{FeaturePerformanceRegression}
\usage{
FeaturePerformanceRegression(
  mmo,
  target,
  phenotype,
  groups,
  model = "lmm",
  normalization = "Z",
  outdir = "FeaturePerformanceRegression",
  width = 6,
  height = 6,
  save_output = TRUE
)
}
\arguments{
\item{mmo}{The mmo object with feature data and metadata}

\item{target}{The name of the feature to analyze}

\item{phenotype}{The name of the phenotype performance in the metadata}

\item{groups}{A vector of group names from the metadata containing performance data}

\item{model}{The type of regression model to use. Options are 'lmm' for linear mixed model, 'lm' for simple linear regression, or 'pearson' for Pearson correlation (default: 'lmm')}

\item{normalization}{The normalization method to use for feature data. Options are 'None', 'Log', 'Meancentered', or 'Z' (default: 'Z')}

\item{outdir}{The directory to save the output files}

\item{width}{The width of the output plot in inches (defa ult: 8)}

\item{height}{The height of the output plot in inches (default: 12)}

\item{save_output}{A logical value indicating whether to save the output plot (default: TRUE)}
}
\value{
A list of the plot and the raw data
}
\description{
This function performs regression analysis of a specific feature against a phenotype performance in the metadata.
It can use linear mixed models (LMM), simple linear regression (LM), or Pearson correlation.
}
