% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{FeaturePresence}
\alias{FeaturePresence}
\title{Convert feature abundances to presence / absence}
\usage{
FeaturePresence(mmo, threshold = 1)
}
\arguments{
\item{mmo}{The mmo object}

\item{threshold}{Numeric threshold for presence (default = 1).
Values > threshold are set to 1, values <= threshold or NA are set to 0.}
}
\value{
The mmo object with a new presence/absence table
stored in mmo$feature_presence
}
\description{
This function converts the feature abundance matrix in an mmo object
into a binary presence/absence matrix and stores it as a new component
of the mmo object (mmo$feature_presence).
}
\details{
A feature is considered present (1) if its abundance is greater than
a specified threshold, and absent (0) otherwise.

This function does NOT overwrite mmo$feature_data.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mmo <- FeaturePresence(mmo, threshold = 1)
\dontshow{\}) # examplesIf}
}
