% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{FeatureToID}
\alias{FeatureToID}
\title{Convert feature names to IDs in the mmo object}
\usage{
FeatureToID(mmo, feature_names)
}
\arguments{
\item{mmo}{The mmo object}

\item{feature_names}{A vector of feature names to convert}
}
\value{
A vector of feature IDs corresponding to the input feature names
}
\description{
This function converts feature names to their corresponding IDs in the mmo object.
}
