% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{GetAlphaDiversity}
\alias{GetAlphaDiversity}
\title{GetAlphaDiversity}
\usage{
GetAlphaDiversity(
  mmo,
  q = 1,
  normalization = "None",
  mode = "weighted",
  distance = "dreams",
  filter_feature = FALSE,
  feature_list = NULL
)
}
\arguments{
\item{mmo}{The mmo object containing feature data and metadata}

\item{q}{The order of the Hill number to calculate (default: 1)}

\item{normalization}{The normalization method to use for feature data. Options are 'None', 'Log', 'Meancentered', or 'Z' (default: 'None')}

\item{mode}{The mode of diversity calculation. Options are 'weighted' or 'unweighted' for chemical distance(default: 'weighted')}

\item{distance}{The distance metric to use for calculating dissimilarity. Options are 'dreams', 'm2ds', or 'cosine' (default: 'dreams')}

\item{filter_feature}{A boolean indicating whether to filter the feature data by a specific list (default: FALSE)}

\item{feature_list}{A list of feature names to filter the feature data by, if filter_feature is TRUE (default: NULL)}
}
\value{
A data frame containing the alpha diversity for each group in the metadata, with columns for group and alpha diversity value.
}
\description{
This function calculates the alpha diversity for a given mmo object, order of the Hill number, normalization method, mode (weighted or unweighted), distance metric, and optional feature filtering.
Unweighted mode uses Hill numbers without considering feature dissimilarity, while weighted mode uses functional Hill numbers that account for feature dissimilarity.
}
\examples{
alpha_diversity <- GetAlphaDiversity(mmo, q = 1, normalization = 'None', 
 mode = 'weighted', distance = 'dreams', filter_feature = FALSE)
alpha_diversity <- GetAlphaDiversity(mmo, q = 2, normalization = 'Z', 
 mode = 'unweighted', filter_feature = TRUE, feature_list = Glucosinolates)
}
