% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{GetDAMs}
\alias{GetDAMs}
\title{Get lists of DAMs (Differentially Accumulated Metabolites) for each comparison}
\usage{
GetDAMs(mmo, fc_cutoff = 0.5849625, pval_cutoff = 0.05)
}
\arguments{
\item{mmo}{The mmo object with pairwise comparison matrix}

\item{fc_cutoff}{The threshold of log2 fold change to be considered significant (default: 0.5849625, which is log2(1.5))}

\item{pval_cutoff}{The threshold of adjusted p-value to be considered significant (default: 0.05)}
}
\value{
A list containing two lists: DAMs_up and DAMs_down
}
\description{
This function generates lists of upregulated and downregulated DAMs for each pairwise comparison in the mmo object.
}
