% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{GetDistanceMat}
\alias{GetDistanceMat}
\title{Get the distance matrix from the mmo object based on the specified distance metric}
\usage{
GetDistanceMat(mmo, distance = "dreams")
}
\arguments{
\item{mmo}{The mmo object}

\item{distance}{The distance metric to use. Options are 'dreams', 'cosine', or 'm2ds'}
}
\value{
The distance matrix corresponding to the specified distance metric
}
\description{
This function retrieves the distance matrix from the mmo object based on the specified distance metric.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
distance_matrix <- GetDistanceMat(mmo, distance = 'dreams')
\dontshow{\}) # examplesIf}
}
