% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{GetFaithPD}
\alias{GetFaithPD}
\title{GetFaithPD}
\usage{
GetFaithPD(feature, metadata, distance_matrix, threshold = 0)
}
\arguments{
\item{feature}{Feature table with columns: id, feature, then sample columns}

\item{metadata}{Metadata table with sample and group columns}

\item{distance_matrix}{Feature distance matrix}

\item{threshold}{Numeric; detection threshold for presence (default: 0)}
}
\value{
A data frame containing the Faith's phylogenetic diversity for each group in the metadata, with columns for group and PD.
}
\description{
This function calculates the Faith's phylogenetic diversity for a given mmo object and distance metric,
to calculate chemically-informed richness
}
