% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{GetFaithPD}
\alias{GetFaithPD}
\title{GetFaithPD}
\usage{
GetFaithPD(mmo, distance = "dreams")
}
\arguments{
\item{mmo}{The mmo object containing feature data and metadata}

\item{distance}{The distance metric to use for feature data. Options are 'dreams', 'jaccard', 'bray', 'unifrac', 'wunifrac' (default: 'dreams')}
}
\value{
A data frame containing the Faith's phylogenetic diversity for each group in the metadata, with columns for group and PD.
}
\description{
This function calculates the Faith's phylogenetic diversity for a given mmo object and distance metric,
to calculate chemically-informed richness
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
faith_pd <- GetFaithPD(mmo, distance = 'dreams')
\dontshow{\}) # examplesIf}
}
