% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{GetFunctionalHillNumber}
\alias{GetFunctionalHillNumber}
\title{GetFunctionalHillNumber}
\usage{
GetFunctionalHillNumber(
  mmo,
  normalization = "None",
  q = 1,
  distance = "dreams"
)
}
\arguments{
\item{mmo}{The mmo object containing feature data and metadata}

\item{normalization}{The normalization method to use for feature data.
Options are 'None', 'Log', 'Meancentered', or 'Z' (default: 'None')}

\item{q}{The order of the Hill number to calculate (default: 1).
Larger q values give more weight to evenness portion of the hill number over richness.}

\item{distance}{The distance metric to use for calculating dissimilarity.
Options are 'dreams', 'm2ds', or 'cosine' (default: 'dreams')}
}
\value{
A data frame containing the functional Hill number for each group in the metadata, with columns for group and hill number.
}
\description{
This function calculates the functional Hill number for a given mmo object, normalization method, and distance metric.
See https://nph.onlinelibrary.wiley.com/doi/full/10.1111/nph.18685 for details of the functional Hill number calculation.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
hill_number <- GetFunctionalHillNumber(mmo,
                                       normalization = 'Z',
                                       q = 1, distance = 'dreams',
                                       filter_id = FALSE)
hill_number <- GetFunctionalHillNumber(mmo, normalization = 'Z',
                                       q = 3, distance = 'dreams',
                                       filter_id = TRUE,
                                       id_list = Glucosinolates)
\dontshow{\}) # examplesIf}
}
