% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{GetFunctionalHillNumber}
\alias{GetFunctionalHillNumber}
\title{GetFunctionalHillNumber}
\usage{
GetFunctionalHillNumber(
  mmo,
  normalization = "None",
  q = 1,
  distance = "dreams",
  filter_feature = FALSE,
  feature_list = NULL
)
}
\arguments{
\item{mmo}{The mmo object containing feature data and metadata}

\item{normalization}{The normalization method to use for feature data. Options are 'None', 'Log', 'Meancentered', or 'Z' (default: 'None')}

\item{q}{The order of the Hill number to calculate (default: 1)}

\item{distance}{The distance metric to use for calculating dissimilarity. Options are 'dreams', 'm2ds', or 'cosine' (default: 'dreams')}

\item{filter_feature}{A boolean indicating whether to filter the feature data by a specific list (default: FALSE)}

\item{feature_list}{A list of feature names to filter the feature data by, if filter_feature is TRUE (default: NULL)}
}
\value{
A data frame containing the functional Hill number for each group in the metadata, with columns for group and hill number.
}
\description{
This function calculates the functional Hill number for a given mmo object, normalization method, and distance metric.
}
