% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{GetGroupMeans}
\alias{GetGroupMeans}
\title{Calculate group means from the mmo object}
\usage{
GetGroupMeans(
  mmo,
  normalization = "None",
  filter_feature = FALSE,
  feature_list = NULL,
  filter_group = FALSE,
  group_list = NULL
)
}
\arguments{
\item{mmo}{The mmo object}

\item{normalization}{The normalization method to use. Options are 'None', 'Log', 'Meancentered', or 'Z'}

\item{filter_feature}{Boolean to filter features based on a provided list (default: FALSE)}

\item{feature_list}{A vector of feature names to filter (default: NULL)}

\item{filter_group}{Boolean to filter groups based on a provided list (default: FALSE)}

\item{group_list}{A vector of group names to filter (default: NULL)}
}
\value{
A data frame containing the mean feature values for each group
}
\description{
This function calculates the mean feature values for each group in the mmo object.
}
