% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{GetNormFeature}
\alias{GetNormFeature}
\title{Get feature data from the mmo object, with normalization options}
\usage{
GetNormFeature(mmo, normalization = "None")
}
\arguments{
\item{mmo}{The mmo object}

\item{normalization}{The normalization method to use. Options are 'None', 'Log', 'Meancentered', or 'Z'}
}
\value{
The feature data corresponding to the specified normalization method
}
\description{
This function retrieves the feature data from the mmo object based on the specified normalization method.
}
