% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{GetPerformanceFeatureCorrelation}
\alias{GetPerformanceFeatureCorrelation}
\title{GetPerformanceFeatureCorrelation}
\usage{
GetPerformanceFeatureCorrelation(
  mmo,
  phenotype,
  groups,
  DAM.list,
  comparisons,
  cor_method = "pearson",
  normalization = "None"
)
}
\arguments{
\item{mmo}{The mmo object with feature data and metadata}

\item{phenotype}{The name of the phenotype performance in the metadata}

\item{groups}{A vector of group names from the metadata containing performance data}

\item{DAM.list}{A list of DAMs to tag features}

\item{comparisons}{A list of pairwise comparisons to add fold change columns}

\item{cor_method}{The correlation method to use. Options are 'pearson', 'spearman', or 'kendall' (default: 'pearson')}

\item{normalization}{The normalization method to use for feature data. Options are 'None', 'Log', 'Meancentered', or 'Z' (default: 'None')}
}
\value{
A data frame containing correlation results for each feature, including effect size (correlation coefficient), p-value, and fold change columns for specified comparisons.
}
\description{
This function calculates the Pearson correlation between each feature and a specified phenotype performance in the metadata.
}
