% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{GetSpecializationIndex}
\alias{GetSpecializationIndex}
\title{GetSpecializationIndex}
\usage{
GetSpecializationIndex(
  mmo,
  normalization = "None",
  filter_group = FALSE,
  group_list = NULL,
  filter_feature = FALSE,
  feature_list = NULL
)
}
\arguments{
\item{mmo}{The mmo object containing feature data and metadata}

\item{normalization}{The normalization method to use for feature data. Options are 'None', 'Log', 'Meancentered', or 'Z' (default: 'None')}

\item{filter_group}{A boolean indicating whether to filter the feature data by a specific group list (default: FALSE)}

\item{group_list}{A list of groups to filter the feature data by, if filter_group is TRUE (default: NULL)}

\item{filter_feature}{A boolean indicating whether to filter the feature data by a specific list (default: FALSE)}

\item{feature_list}{A list of feature names to filter the feature data by, if filter_feature is TRUE (default: NULL)}
}
\value{
A data frame containing the specialization index for each group in the metadata, with columns for group and specialization index.
}
\description{
This function calculates the specialization index for a given mmo object, normalization method, and optional filtering by groups and features.
}
