% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{IDToFeature}
\alias{IDToFeature}
\title{Convert feature IDs to names in the mmo object}
\usage{
IDToFeature(mmo, feature_ids)
}
\arguments{
\item{mmo}{The mmo object}

\item{feature_ids}{A vector of feature IDs to convert}
}
\value{
A vector of feature names corresponding to the input feature IDs
}
\description{
This function converts feature IDs to their corresponding names in the mmo object.
}
