% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{IDToFeature}
\alias{IDToFeature}
\title{Convert feature IDs to names in the mmo object}
\usage{
IDToFeature(mmo, feature_ids)
}
\arguments{
\item{mmo}{The mmo object}

\item{feature_ids}{A vector of feature IDs to convert}
}
\value{
A vector of feature names corresponding to the input feature IDs
}
\description{
This function converts feature IDs to their corresponding names in the mmo object.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
feature_names <- IDToFeature(mmo, feature_ids = c("1219", "2250", "3360"))
feature_names <- IDToFeature(mmo, feature_ids = mmo$feature_data$id[1:10])
feature_names <- IDToFeature(mmo,
 feature_ids = FeatureToID(mmo, feature_names = Glucosinolates)
) # if Glucosinolates is a vector of feature names
\dontshow{\}) # examplesIf}
}
