% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{LogNormalization}
\alias{LogNormalization}
\title{Log-normalize the peak area in the mmo object}
\usage{
LogNormalization(mmo)
}
\arguments{
\item{mmo}{The mmo object}
}
\value{
The mmo object with log-normalized feature data (mmo$log)
}
\description{
This function applies log2 transformation to the peak area in the feature data of the mmo object.
Run ReplaceZero() before this function to avoid -Inf values.
}
\examples{
mmo <- LogNormalization(mmo)
}
