% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{MassNormalization}
\alias{MassNormalization}
\title{Use sample mass in the metadata file to normalize the peak area}
\usage{
MassNormalization(mmo)
}
\arguments{
\item{mmo}{The mmo object}
}
\value{
The mmo object with normalized feature data (mmo$feature_data)
}
\description{
This function normalizes the peak area in the feature data of the mmo object by the mass of each sample, provided in the metadata.
The feature data is replaced by (original value * mean mass) / sample mass.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
mmo <- MassNormalization(mmo)
\dontshow{\}) # examplesIf}
}
