% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{MeancenterNormalization}
\alias{MeancenterNormalization}
\title{Mean-center the peak area in the mmo object}
\usage{
MeancenterNormalization(mmo)
}
\arguments{
\item{mmo}{The mmo object}
}
\value{
The mmo object with mean-centered feature data (mmo$meancentered)
}
\description{
This function applies mean-centering to the peak area in the feature data of the mmo object.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
mmo <- MeancenterNormalization(mmo)
\dontshow{\}) # examplesIf}
}
