% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{MSEA}
\alias{MSEA}
\title{Metabolite Set Enrichment Analysis (MSEA)}
\usage{
MSEA(
  mmo,
  feature_name,
  feature_score,
  term_level = "NPC_class",
  pthr = 0.05,
  outdir = "MSEA",
  width = 8,
  height = 12,
  sig = FALSE,
  save_output = TRUE
)
}
\arguments{
\item{mmo}{The mmo object with sirius annotation and normalized data}

\item{feature_name}{A vector of feature names corresponding to the feature scores}

\item{feature_score}{A vector of feature scores (e.g., log2 fold changes)}

\item{term_level}{The level of term to use for enrichment analysis.
Options are 'NPC_pathway', 'NPC_superclass', 'NPC_class',
'ClassyFire_superclass', 'ClassyFire_class', 'ClassyFire_subclass',
'ClassyFire_level5', or 'ClassyFire_most_specific' (default: 'NPC_class')}

\item{pthr}{The threshold for adjusted p-value to be considered significant (default: 0.05)}

\item{outdir}{The directory to save the output files (default: 'MSEA')}

\item{width}{The width of the output plot in inches (default: 8)}

\item{height}{The height of the output plot in inches (default: 12)}

\item{sig}{A logical value indicating whether to return only significant terms (default: FALSE)}

\item{save_output}{A logical value indicating whether to save the output plot (default: TRUE)}
}
\value{
A list of the plot and the enrichment results
}
\description{
This function performs Metabolite Set Enrichment Analysis (MSEA) using the fgsea package.
It takes a ranked list of feature scores and tests for enrichment of metabolite sets based on Canopus-predicted terms.
The results are saved as a CSV file and a PDF plot.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Perform MSEA using NPC_class level
MSEA(
 mmo, feature_name = rownames(DE_results), feature_score = DE_results$log2FoldChange,
 term_level = 'NPC_class', pthr = 0.05, outdir = 'MSEA_NPC_class',
 width = 8, height = 12, sig = FALSE
)
\dontshow{\}) # examplesIf}
}
