% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{PairwiseComp}
\alias{PairwiseComp}
\title{Perform pairwise comparison between two groups in the mmo object}
\usage{
PairwiseComp(mmo, group1, group2, correction = "BH")
}
\arguments{
\item{mmo}{The mmo object}

\item{group1}{The name of the nominator group}

\item{group2}{The name of the denominator group}

\item{correction}{The method for multiple comparison correction. Options are 'BH', 'holm', 'bonferroni', etc. Inherits from p.adjust()}
}
\description{
This function performs pairwise comparison between two groups in the mmo object,
calculating log2 fold change and adjusted p-values for given comparison of two groups.
The function adds the results to the mmo$pairwise data frame.
}
