% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{PCAplot}
\alias{PCAplot}
\title{Plots PCA and performs PERMANOVA}
\usage{
PCAplot(
  mmo,
  color,
  outdir = "PCA",
  normalization = "Z",
  filter_feature = FALSE,
  feature_list = NULL,
  filter_group = FALSE,
  group_list = NULL,
  label = TRUE
)
}
\arguments{
\item{mmo}{The mmo object with feature data and metadata}

\item{color}{A vector of colors for the groups in the plot. Make sure the names correspond to the group names in metadata}

\item{outdir}{The output file path for the PCA plot (default: 'PCA')}

\item{normalization}{The normalization method to use for feature data. Options are 'None', 'Log', 'Meancentered', or 'Z' (default: 'Z')}

\item{filter_feature}{Boolean to filter features by feature_list (default: FALSE)}

\item{feature_list}{A vector of feature names to filter (default: NULL)}

\item{filter_group}{Boolean to filter groups by group_list (default: FALSE)}

\item{group_list}{A vector of group names to filter (default: NULL)}

\item{label}{Boolean to indicate whether to label points with sample names (default: TRUE)}
}
\description{
This function performs PCA analysis and generates a PCA plot with optional filtering of features and groups.
It also conducts PERMANOVA and saves the results to CSV files.
}
\examples{
PCAplot(
 mmo, color = c("Control" = "blue", "Treatment1" = "red", "Treatment2" = "green"), 
 outdir = 'PCA_plot', normalization = 'None', 
 filter_feature = FALSE, filter_group = FALSE, label = FALSE
)
PCAplot(
 mmo, color = c("Control" = "blue", "Treatment1" = "red"), 
 outdir = 'PCA_plot', normalization = 'Z', 
 filter_feature = TRUE, feature_list = Glucosinolates, 
 filter_group = TRUE, group_list = c("Control", "Treatment1"), label = TRUE
)
}
