% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{PCoAplot}
\alias{PCoAplot}
\title{PCoAplot}
\usage{
PCoAplot(
  mmo,
  betadiv,
  outdir,
  width = 6,
  height = 6,
  color,
  save_output = TRUE
)
}
\arguments{
\item{mmo}{The mmo object containing metadata}

\item{betadiv}{The beta diversity distance matrix, output of GetBetaDiversity}

\item{outdir}{The prefix for the output files}

\item{width}{The width of the output PCoA plot (default: 6}

\item{height}{The height of the output PCoA plot (default: 6)}

\item{color}{A vector of colors for the points in the plot}

\item{save_output}{A logical value indicating whether to save the output plot (default: TRUE)}
}
\value{
A list containing the PCoA plot, the PCoA coordinates, and the PERMANOVA results
}
\description{
This function generates a Principal Coordinates Analysis (PCoA) plot based on the provided beta diversity distance matrix.
It also performs PERMANOVA analysis to assess the significance of group differences and saves the
results to CSV files.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
beta_diversity <- GetBetaDiversity(mmo, method = 'Gen.Uni',
 normalization = 'None', distance = 'dreams', filter_id = FALSE)
PCoAplot(mmo, betadiv = beta_diversity, outdir = 'output/PCoA', width = 6, height = 6)
\dontshow{\}) # examplesIf}
}
