% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{PlotNPCStackedBar}
\alias{PlotNPCStackedBar}
\title{PlotNPCStackedBar}
\usage{
PlotNPCStackedBar(mmo, group_col, output_file, width = 6, height = 3)
}
\arguments{
\item{mmo}{The mmo object with sirius annotation and normalized data}

\item{group_col}{The column name in metadata to use for grouping samples}

\item{output_file}{The output file path for the stacked bar plot (e.g., 'NPC_stacked_bar.png')}

\item{width}{The width of the output plot}

\item{height}{The height of the output plot}
}
\description{
This function generates a stacked bar plot showing the count of features in each group categorized by NPC_pathway.
It uses the mmo object with sirius annotation and normalized data.
Make sure you don't run ReplaceZero() before using this function, as it may remove presence/absence information.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
PlotNPCStackedBar(
 mmo, group_col = 'treatment', 
 output_file = 'NPC_stacked_bar.png', width = 6, height = 3
)
\dontshow{\}) # examplesIf}
}
