% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V3.R
\name{ReplaceZero}
\alias{ReplaceZero}
\title{#' Replace zero and NA values in the mmo object}
\usage{
ReplaceZero(mmo, method = c("one", "half_min"))
}
\arguments{
\item{mmo}{An mmo object containing \code{feature_data}}

\item{method}{Replacement method:
\itemize{
\item "one": replace zeros and NA values with 1
\item "half_min": replace zeros and NA values with half of the smallest
non-zero value in each feature (row)
}}
}
\value{
The updated mmo object
}
\description{
This function replaces zero values in the feature table of an mmo object.
Imputed data are stored in mmo$imputed_feature_data, to be used for
downsteam analyses including PairwiseComp().
Note that imputation affects normalizations (Log-transformation, etc.),
as well as chemical diversity calculations that uses presence/absence.
}
