% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{ReplaceZero}
\alias{ReplaceZero}
\title{Replace zero and NA values in the mmo object}
\usage{
ReplaceZero(mmo, method = "one")
}
\arguments{
\item{mmo}{The mmo object}

\item{method}{The method to use for replacement. Options are 'one' (replace with 1) or 'half_mean' (replace with half of the smallest non-zero value in the row)}
}
\description{
This function replaces zero and NA values in the feature data of the mmo object.
}
