% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{SwitchGroup}
\alias{SwitchGroup}
\title{Switch the group column in the mmo object}
\usage{
SwitchGroup(mmo, new_group_col)
}
\arguments{
\item{mmo}{The mmo object}

\item{new_group_col}{The name of the new group column in the metadata file}
}
\value{
The mmo object with the updated group column
}
\description{
This function switches the group column in the metadata of the mmo object to a new specified column.
The new group column must exist in the metadata file.
}
