% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{VolcanoPlot}
\alias{VolcanoPlot}
\title{Volcano plot for visualizing differential metabolite analysis results}
\usage{
VolcanoPlot(
  mmo,
  comp,
  topk = 10,
  pthr = 0.05,
  outdir = "volcano.png",
  height = 5,
  width = 5
)
}
\arguments{
\item{mmo}{The mmo object with pairwise comparison matrix}

\item{comp}{The comparison to visualize, e.g., 'group1_vs_group2}

\item{topk}{The number of top features to label in the plot (default: 10)}

\item{pthr}{The threshold of adjusted p-value to be considered significant (default: 0.05)}

\item{outdir}{The output file path for the volcano plot (default: 'volcano.png')}

\item{height}{The height of the output plot in inches (default: 5)}

\item{width}{The width of the output plot in inches (default: 5)}
}
\description{
This function generates a volcano plot using data from mmo$pairwise,
highlighting upregulated and downregulated features based on log2 fold change and adjusted p-value
}
