% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2_MC.R
\name{ZNormalization}
\alias{ZNormalization}
\title{Z-normalize the peak area in the mmo object}
\usage{
ZNormalization(mmo)
}
\arguments{
\item{mmo}{The mmo object}
}
\value{
The mmo object with Z-normalized feature data (mmo$zscore)
}
\description{
This function applies Z-score normalization to the peak area in the feature data of the mmo object.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mmo <- ZNormalization(mmo)
\dontshow{\}) # examplesIf}
}
